Revert "fixed image upload to gpu"
This reverts commit 412aff7383a5252107dae45c693d6d60811395af.

+++ OptixPathtracer/source/ModelLoading/Texture.h
@@ -8,6 +8,6 @@ struct Texture {
         if (pixel) delete[] pixel;
     }
 
-    float* pixel{ nullptr };
+    uint32_t* pixel{ nullptr };
     glm::ivec2 resolution{ -1 };
 };
